Restructure app and move modules into their own folder
