+++ extension/client/chat-handlers.js
