+++ extension/client/index.js
