+++ extension/index.js
@@ -17,13 +17,17 @@ module.exports = (nodecg) => {
   }
 
   context.nodecg = nodecg
-  context.twitch = require('./twitch')
+  context.twitch = require('./client')
+
+  // mount our refresh route under the main nodecg express app
+  const service = require('./service')
+  nodecg.mount(service)
 
   const api = require('./api')
-  require('./service')
-  require('./eventLog')
-  require('./channel')
-  require('./chat')
+
+  require('./modules/eventLog')
+  require('./modules/channel')
+  require('./modules/chat')
 
   // exposes the module's EventEmitter, and proxies requests to the current
   // twitch instance if one exists--we can't just return the twitch object,
