+++ extension/modules/channel/index.js
