+++ extension/modules/chat/index.js
